# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for The Agency Engine
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  clerkId: string;
  email: string;
  /** @nullable */
  displayName?: string | null;
  gaBalance: number;
  chsScore: number;
  createdAt: string;
}

export type DashboardStatsChsTrend = typeof DashboardStatsChsTrend[keyof typeof DashboardStatsChsTrend];


export const DashboardStatsChsTrend = {
  up: 'up',
  down: 'down',
  stable: 'stable',
} as const;

export interface DashboardStats {
  gaBalance: number;
  gaEarnedToday: number;
  gaSpentToday: number;
  chsScore: number;
  chsTrend: DashboardStatsChsTrend;
  tasksCompletedTotal: number;
  humanTasksCompletedTotal: number;
  streakDays: number;
}

export interface GaBalance {
  balance: number;
  dailyAllowance: number;
  lastResetAt: string;
}

export interface GaLedgerEntry {
  id: number;
  delta: number;
  reason: string;
  createdAt: string;
}

export interface SpendGaRequest {
  /** @minimum 1 */
  amount: number;
  reason: string;
}

export interface EarnGaRequest {
  /** ID of a passing creative-labor or human-task submission. Single-use — once credited this call returns 400. */
  submissionId: number;
}

export interface RecordChsRequest {
  /**
     * The interaction text to score (a prompt, submission, or reflection).
     * @minLength 5
     */
  text: string;
  /** Optional label prepended to the text when scoring (e.g. task title). */
  context?: string;
}

export type ChsScoreBand = typeof ChsScoreBand[keyof typeof ChsScoreBand];


export const ChsScoreBand = {
  critical: 'critical',
  poor: 'poor',
  fair: 'fair',
  good: 'good',
  thriving: 'thriving',
} as const;

export interface ChsFactors {
  complexity: number;
  originality: number;
  depth: number;
  effort: number;
}

export interface ChsScore {
  score: number;
  band: ChsScoreBand;
  factors: ChsFactors;
  updatedAt: string;
}

export interface ChsHistoryPoint {
  date: string;
  averageScore: number;
}

export interface GatewayTaskInfo {
  type: string;
  title: string;
  description: string;
  instructions: string;
  placeholder: string;
  gaBonus: number;
  /** Countdown duration in seconds, or null if not a timed task */
  timedSeconds: number | null;
}

export interface GatewayStatus {
  locked: boolean;
  currentTask: GatewayTaskInfo | null;
  unlockExpiresAt: string | null;
  minutesRemaining: number | null;
  sessionSpend: number;
  sessionSpendThreshold: number;
  sessionMinutesElapsed: number;
  sessionMinutesLimit: number;
}

export interface GatewaySubmitRequest {
  /**
     * The user's response to the assigned human task
     * @minLength 20
     */
  response: string;
}

export interface GatewaySubmitResult {
  passed: boolean;
  verdict: string;
  gaBonus: number;
  newBalance: number;
  unlockExpiresAt: string | null;
}

export interface CreativeLaborTask {
  id: number;
  slug: string;
  type: string;
  title: string;
  description: string;
  instructions: string;
  gaReward: number;
}

export interface SubmitCreativeLaborRequest {
  taskSlug: string;
  /** @minLength 10 */
  content: string;
}

export interface CreativeLaborVerdict {
  passed: boolean;
  qualityNotes: string;
  gaRewarded: number;
  currentBalance: number;
  submissionId: number;
}