# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for The Agency Engine
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: user
    description: User profile and stats
  - name: ga
    description: Genius Answer token management
  - name: chs
    description: Cognitive Health Score
  - name: creative-labor
    description: Creative labor tasks and AI evaluation
  - name: gateway
    description: Proof of Humanity Gateway — circuit-breaker for AI access
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /me:
    get:
      operationId: getMe
      tags: [user]
      summary: Get current user profile
      description: Returns user profile with current GA token balance
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized

  /stats:
    get:
      operationId: getStats
      tags: [user]
      summary: Get dashboard stats
      description: Returns summary counts for the dashboard
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          description: Unauthorized

  /ga/balance:
    get:
      operationId: getGaBalance
      tags: [ga]
      summary: Get GA token balance
      description: Returns the current GA token balance for the user
      responses:
        "200":
          description: GA token balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GaBalance"
        "401":
          description: Unauthorized

  /ga/history:
    get:
      operationId: getGaHistory
      tags: [ga]
      summary: Get GA token history
      description: Returns the ledger of GA token transactions
      responses:
        "200":
          description: GA token history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GaLedgerEntry"
        "401":
          description: Unauthorized

  /ga/spend:
    post:
      operationId: spendGa
      tags: [ga]
      summary: Spend GA tokens
      description: Deducts GA tokens from the user's balance for an AI query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpendGaRequest"
      responses:
        "200":
          description: Updated GA balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GaBalance"
        "400":
          description: Insufficient balance or invalid amount
        "401":
          description: Unauthorized

  /ga/earn:
    post:
      operationId: earnGa
      tags: [ga]
      summary: Claim earned GA tokens
      description: |
        Credits GA tokens to the user's balance for a verified creative-labor or
        human-task submission. Requires a `submissionId` from a passing evaluation —
        each submission can only be credited once. This prevents self-award exploits
        while keeping token issuance in a single trusted path.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EarnGaRequest"
      responses:
        "200":
          description: Updated GA balance after crediting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GaBalance"
        "400":
          description: Invalid submissionId or already credited
        "401":
          description: Unauthorized

  /chs/record:
    post:
      operationId: recordChs
      tags: [chs]
      summary: Record a cognitive interaction and update CHS
      description: |
        Scores the provided interaction text across 4 dimensions (complexity, originality,
        depth, effort) using deterministic heuristics, then updates the user's CHS via
        EWMA blending (weight 0.3). Call this whenever the user writes or submits content.
        Creative-labor submissions automatically trigger this internally on the server.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordChsRequest"
      responses:
        "200":
          description: Updated CHS after recording the interaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChsScore"
        "400":
          description: Invalid request (text too short)
        "401":
          description: Unauthorized

  /chs/current:
    get:
      operationId: getChsCurrent
      tags: [chs]
      summary: Get current Cognitive Health Score
      description: Returns the user's current CHS and factor breakdown
      responses:
        "200":
          description: Current CHS
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChsScore"
        "401":
          description: Unauthorized

  /chs/history:
    get:
      operationId: getChsHistory
      tags: [chs]
      summary: Get CHS history
      description: Returns daily CHS averages for the past 30 days
      responses:
        "200":
          description: CHS history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChsHistoryPoint"
        "401":
          description: Unauthorized

  /gateway/status:
    get:
      operationId: getGatewayStatus
      tags: [gateway]
      summary: Get current gateway state
      description: |
        Returns whether the gateway is locked, the assigned human task (if locked),
        remaining access time (if unlocked), and session spend/time progress.
        Auto-runs expiry and threshold checks so the response always reflects live state.
      responses:
        "200":
          description: Current gateway status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayStatus"
        "401":
          description: Unauthorized

  /gateway/trigger:
    post:
      operationId: triggerGateway
      tags: [gateway]
      summary: Manually lock the gateway
      description: Locks the gateway and assigns a random human task. No-op if already locked.
      responses:
        "200":
          description: Updated gateway status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayStatus"
        "401":
          description: Unauthorized

  /gateway/submit:
    post:
      operationId: submitGateway
      tags: [gateway]
      summary: Submit a human-task response
      description: |
        Evaluates the user's response to the assigned human task via LLM.
        On pass: unlocks the gateway for 30 minutes and awards a GA bonus.
        On fail: returns feedback but leaves the gateway locked.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GatewaySubmitRequest"
      responses:
        "200":
          description: Evaluation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewaySubmitResult"
        "400":
          description: Gateway not locked, no active task, or response too short
        "401":
          description: Unauthorized

  /creative-labor/tasks:
    get:
      operationId: getCreativeLaborTasks
      tags: [creative-labor]
      summary: Get available creative labor tasks
      description: Returns the list of task templates available for completion
      responses:
        "200":
          description: List of creative labor tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CreativeLaborTask"
        "401":
          description: Unauthorized

  /creative-labor/submit:
    post:
      operationId: submitCreativeLabor
      tags: [creative-labor]
      summary: Submit creative labor for AI evaluation
      description: |
        Submits creative work for AI evaluation. Returns a verdict with a submissionId.
        If the verdict passes, call POST /ga/earn with the submissionId to claim tokens.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitCreativeLaborRequest"
      responses:
        "200":
          description: Evaluation verdict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreativeLaborVerdict"
        "400":
          description: Invalid submission
        "401":
          description: Unauthorized

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    UserProfile:
      type: object
      properties:
        clerkId:
          type: string
        email:
          type: string
        displayName:
          type: ["string", "null"]
        gaBalance:
          type: integer
        chsScore:
          type: integer
        createdAt:
          type: string
      required:
        - clerkId
        - email
        - gaBalance
        - chsScore
        - createdAt

    DashboardStats:
      type: object
      properties:
        gaBalance:
          type: integer
        gaEarnedToday:
          type: integer
        gaSpentToday:
          type: integer
        chsScore:
          type: integer
        chsTrend:
          type: string
          enum: [up, down, stable]
        tasksCompletedTotal:
          type: integer
        humanTasksCompletedTotal:
          type: integer
        streakDays:
          type: integer
      required:
        - gaBalance
        - gaEarnedToday
        - gaSpentToday
        - chsScore
        - chsTrend
        - tasksCompletedTotal
        - humanTasksCompletedTotal
        - streakDays

    GaBalance:
      type: object
      properties:
        balance:
          type: integer
        dailyAllowance:
          type: integer
        lastResetAt:
          type: string
      required:
        - balance
        - dailyAllowance
        - lastResetAt

    GaLedgerEntry:
      type: object
      properties:
        id:
          type: integer
        delta:
          type: integer
        reason:
          type: string
        createdAt:
          type: string
      required:
        - id
        - delta
        - reason
        - createdAt

    SpendGaRequest:
      type: object
      properties:
        amount:
          type: integer
          minimum: 1
        reason:
          type: string
      required:
        - amount
        - reason

    EarnGaRequest:
      type: object
      properties:
        submissionId:
          type: integer
          description: ID of a passing creative-labor or human-task submission. Single-use — once credited this call returns 400.
      required:
        - submissionId

    RecordChsRequest:
      type: object
      properties:
        text:
          type: string
          minLength: 5
          description: The interaction text to score (a prompt, submission, or reflection).
        context:
          type: string
          description: Optional label prepended to the text when scoring (e.g. task title).
      required:
        - text

    ChsScore:
      type: object
      properties:
        score:
          type: integer
        band:
          type: string
          enum: [critical, poor, fair, good, thriving]
        factors:
          $ref: "#/components/schemas/ChsFactors"
        updatedAt:
          type: string
      required:
        - score
        - band
        - factors
        - updatedAt

    ChsFactors:
      type: object
      properties:
        complexity:
          type: integer
        originality:
          type: integer
        depth:
          type: integer
        effort:
          type: integer
      required:
        - complexity
        - originality
        - depth
        - effort

    ChsHistoryPoint:
      type: object
      properties:
        date:
          type: string
        averageScore:
          type: integer
      required:
        - date
        - averageScore

    GatewayTaskInfo:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        description:
          type: string
        instructions:
          type: string
        placeholder:
          type: string
        gaBonus:
          type: integer
        timedSeconds:
          type: integer
          nullable: true
          description: Countdown duration in seconds, or null if not a timed task
      required:
        - type
        - title
        - description
        - instructions
        - placeholder
        - gaBonus
        - timedSeconds

    GatewayStatus:
      type: object
      properties:
        locked:
          type: boolean
        currentTask:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/GatewayTaskInfo"
        unlockExpiresAt:
          type: string
          nullable: true
        minutesRemaining:
          type: integer
          nullable: true
        sessionSpend:
          type: integer
        sessionSpendThreshold:
          type: integer
        sessionMinutesElapsed:
          type: integer
        sessionMinutesLimit:
          type: integer
      required:
        - locked
        - currentTask
        - unlockExpiresAt
        - minutesRemaining
        - sessionSpend
        - sessionSpendThreshold
        - sessionMinutesElapsed
        - sessionMinutesLimit

    GatewaySubmitRequest:
      type: object
      properties:
        response:
          type: string
          minLength: 20
          description: The user's response to the assigned human task
      required:
        - response

    GatewaySubmitResult:
      type: object
      properties:
        passed:
          type: boolean
        verdict:
          type: string
        gaBonus:
          type: integer
        newBalance:
          type: integer
        unlockExpiresAt:
          type: string
          nullable: true
      required:
        - passed
        - verdict
        - gaBonus
        - newBalance
        - unlockExpiresAt

    CreativeLaborTask:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        type:
          type: string
        title:
          type: string
        description:
          type: string
        instructions:
          type: string
        gaReward:
          type: integer
      required:
        - id
        - slug
        - type
        - title
        - description
        - instructions
        - gaReward

    SubmitCreativeLaborRequest:
      type: object
      properties:
        taskSlug:
          type: string
        content:
          type: string
          minLength: 10
      required:
        - taskSlug
        - content

    CreativeLaborVerdict:
      type: object
      properties:
        passed:
          type: boolean
        qualityNotes:
          type: string
        gaRewarded:
          type: integer
        currentBalance:
          type: integer
        submissionId:
          type: integer
      required:
        - passed
        - qualityNotes
        - gaRewarded
        - currentBalance
        - submissionId