# lib/api-zod/src/generated/types/gatewayStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for The Agency Engine
 * OpenAPI spec version: 0.1.0
 */
import type { GatewayTaskInfo } from './gatewayTaskInfo';

export interface GatewayStatus {
  locked: boolean;
  currentTask: GatewayTaskInfo | null;
  unlockExpiresAt: string | null;
  minutesRemaining: number | null;
  sessionSpend: number;
  sessionSpendThreshold: number;
  sessionMinutesElapsed: number;
  sessionMinutesLimit: number;
}