# lib/api-zod/src/generated/types/gatewaySubmitRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for The Agency Engine
 * OpenAPI spec version: 0.1.0
 */

export interface GatewaySubmitRequest {
  /**
     * The user's response to the assigned human task
     * @minLength 20
     */
  response: string;
}