# lib/api-zod/src/generated/types/gatewaySubmitResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for The Agency Engine
 * OpenAPI spec version: 0.1.0
 */

export interface GatewaySubmitResult {
  passed: boolean;
  verdict: string;
  gaBonus: number;
  newBalance: number;
  unlockExpiresAt: string | null;
}