# lib/api-zod/src/generated/types/gatewayTaskInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for The Agency Engine
 * OpenAPI spec version: 0.1.0
 */

export interface GatewayTaskInfo {
  type: string;
  title: string;
  description: string;
  instructions: string;
  placeholder: string;
  gaBonus: number;
  /** Countdown duration in seconds, or null if not a timed task */
  timedSeconds: number | null;
}